+++ common/src/main/java/net/mehvahdjukaar/supplementaries/mixins/RecipeManagerMixin.java
@@ -1,0 +1,43 @@
+package net.mehvahdjukaar.supplementaries.mixins;
+
+import com.google.common.collect.ImmutableMap;
+import com.google.gson.JsonElement;
+import net.mehvahdjukaar.supplementaries.common.items.crafting.RecipeBookHack;
+import net.minecraft.network.PacketListener;
+import net.minecraft.network.protocol.game.ServerboundPlaceRecipePacket;
+import net.minecraft.resources.ResourceLocation;
+import net.minecraft.server.level.ServerPlayer;
+import net.minecraft.server.network.ServerGamePacketListenerImpl;
+import net.minecraft.server.packs.resources.ResourceManager;
+import net.minecraft.util.profiling.ProfilerFiller;
+import net.minecraft.world.inventory.RecipeBookMenu;
+import net.minecraft.world.item.crafting.Recipe;
+import net.minecraft.world.item.crafting.RecipeManager;
+import org.spongepowered.asm.mixin.Final;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.Shadow;
+import org.spongepowered.asm.mixin.injection.At;
+import org.spongepowered.asm.mixin.injection.Inject;
+import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
+import org.spongepowered.asm.mixin.injection.callback.LocalCapture;
+
+import java.util.Map;
+
+@Mixin(RecipeManager.class)
+public abstract class RecipeManagerMixin {
+
+
+    @Inject(method = "apply(Ljava/util/Map;Lnet/minecraft/server/packs/resources/ResourceManager;Lnet/minecraft/util/profiling/ProfilerFiller;)V",
+           locals = LocalCapture.CAPTURE_FAILEXCEPTION,
+            at = @At(value = "INVOKE",
+                    shift = At.Shift.BY,
+                    by = 2,
+                    target = "Lcom/google/common/collect/ImmutableMap;builder()Lcom/google/common/collect/ImmutableMap$Builder;"))
+    public void addSpecialRecipeDisplays(Map<ResourceLocation, JsonElement> object, ResourceManager resourceManager,
+                                         ProfilerFiller profiler, CallbackInfo ci, Map map,
+                                         ImmutableMap.Builder<ResourceLocation, Recipe<?>> builder) {
+
+        var v = RecipeBookHack.createClientRecipes();
+        v.values().forEach(r->r.forEach(h->h.forEach(j->builder.put(j.getId(),j))));
+    }
+}